# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "app para cotizaciones de criptos",
        "title": "Cripto Api",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/cotization": {
            "post": {
                "description": "Save a cryptocurrency along with its quote using the provided name and API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cryptocurrencies"
                ],
                "summary": "Save cryptocurrency with quote",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cryptocurrency name",
                        "name": "nombre",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API for obtaining the quote",
                        "name": "api",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with a message",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
        "/cryptocurrencies/cryptocurrency/{id}": {
            "get": {
                "description": "Retrieve a cryptocurrency by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cryptocurrencies"
                ],
                "summary": "Get cryptocurrency by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cryptocurrency ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/primerProjecto_internal_entities_criptomonedas.CriptoMoneda"
                        }
                    }
                }
            }
        },
...
        "/cryptocurrency/{id}": {
            "put": {
                "description": "Update the details of a cryptocurrency by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cryptocurrencies"
                ],
                "summary": "Update cryptocurrency by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cryptocurrency ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Cryptocurrency Data",
                        "name": "cryptocurrency",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/primerProjecto_internal_entities_criptomonedas.CriptoMoneda"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with a message",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        },
...
        "primerProjecto_internal_entities_criptomonedas.CriptoMoneda": {
            "description": "Estructura que define una criptomoneda.",
            "type": "object",
            "properties": {
                "codigo": {
                    "description": "Codigo es el código de la criptomoneda.\n@example BTC",
                    "type": "string"
                },
                "id": {
                    "description": "ID es el identificador único de la criptomoneda.\n@example 1",
                    "type": "integer"
                },
                "nombre": {
                    "description": "Nombre es el nombre de la criptomoneda.\n@example Bitcoin",
                    "type": "string"
                }
            }
        },
...
    }
}